Revert "Override MediaQuery for WidgetsApp (#81295)"
This reverts commit b9bf5769338d52f33e7d8734c59fb9d7463e6471.
